Extract database connection into it's own file
This allows us to connect to the database without setting up the model.
We need this when we're first creating the database structure.
